actualizacion de funciones
Se actualiza una libreria de funciones y se crea el script para cargar cartera

+++ Carga_solic_v1.py
@@ -11,6 +11,8 @@ import os
 import pymysql
 import warnings
 import datetime
+import sys
+import Complement_functions as cf
 
 #Variables
 path = os.getcwd() #obtiene el directorio de trabajo actual
@@ -19,17 +21,6 @@ files = []
 warnings.simplefilter('error', UserWarning)
 
 #funciones
-def logging_carga(cursor_con, filename, staging_table):
-    mysql_log_load = "insert into Operacion_datamart.Importacion "
-    mysql_log_load += " select curdate() as fecha, '" + filename + "' as nombre_archivo, user() as Usuario, count(*) as Registros from Staging." + staging_table + "; ";
-    cursor_con.execute(mysql_log_load)
-    
-def logging_proceso(cursor_con, process, total_steps, step, descripcion):
-    Etapa = str(step) + "/" + str(total_steps) + ") " + descripcion
-    mysql_log_task = "insert into Operacion_datamart.Logs_procesos (fecha, Proceso, Etapa) "
-    mysql_log_task += " select now() as fecha, '" + process + "' , '" + Etapa + "';"
-    #print(mysql_log_task)
-    cursor_con.execute(mysql_log_task)
     
 
 #Constantes
@@ -70,10 +61,15 @@ if filename in files:
                           autocommit=True,
                           local_infile=1)
         cursor = con.cursor()
+        if cf.Validacion_archivo(cursor, filename):
+            print('Archivo previamente cargado: ' + filename)
+            con.close()
+            sys.exit()
+
         cursor.execute('truncate table Staging.' + staging_table + ';')
         cursor.execute(load_sql)
-        logging_carga(cursor, filename, staging_table)
-        logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Carga archivo solicitudes')
+        cf.logging_carga(cursor, filename, staging_table)
+        cf.logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Carga archivo solicitudes')
 #Staging
         paso = 2
         staging_step_1a = "update Staging." + staging_table + " Set FECHASOL = concat(substr(FECHASOL,7,4), '-', substr(FECHASOL,1,2), '-', substr(FECHASOL,4,2))"
@@ -88,14 +84,14 @@ if filename in files:
         staging_step_1d = "update Staging." + staging_table + " Set FECHA_APERT = concat(substr(FECHA_APERT,7,4), '-', substr(FECHA_APERT,1,2), '-', substr(FECHA_APERT,4,2))"
         staging_step_1d += "where length(FECHA_APERT) = 10;"
         cursor.execute(staging_step_1d)
-        logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Actualiza formato de fechas')
+        cf.logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Actualiza formato de fechas')
 
         paso = 3
         staging_step_2a = "insert into Staging.tmp_solicitudes_rvw select * from Staging.tmp_solicitudes where length(FECHASOL) <> 10;"
         cursor.execute(staging_step_2a)
         staging_step_2b = "delete from Staging." + staging_table + " where length(FECHASOL) <> 10;"
         cursor.execute(staging_step_2b)
-        logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Elimina registros con fechas incorrectas')
+        cf.logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Elimina registros con fechas incorrectas')
 
         paso = 4
         staging_step_3 = "update " + table + " a, Staging." + staging_table + " b set a.statussol = b.statussol,"
@@ -106,7 +102,7 @@ if filename in files:
         staging_step_3 += " a.TEL_CEL=b.TEL_CEL"
         staging_step_3 += " where a.numsolicitud = b.numsolicitud;"
         cursor.execute(staging_step_3)
-        logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Actualiza status solicitudes existentes')
+        cf.logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'Actualiza status solicitudes existentes')
 
         paso = 5
         staging_step_4 = "insert into " + table + " select"
@@ -218,7 +214,7 @@ if filename in files:
         staging_step_4 += " on a.numsolicitud = b.numsolicitud"
         staging_step_4 += " where b.numsolicitud is null and a.tel_cel not in ('BANCO','TIENDA');"
         cursor.execute(staging_step_4)
-        logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'inserta solicitudes nuevas')
+        cf.logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso,paso,'inserta solicitudes nuevas')
 
         paso = 6
         staging_step_5 = "insert into Datos_generales.Domicilio (num_cliente) select distinct a.numcte"
@@ -226,7 +222,7 @@ if filename in files:
         staging_step_5 += " on a.NUMCTE = b.Num_cliente"
         staging_step_5 += " where b.Num_cliente is null;"
         cursor.execute(staging_step_5)
-        logging_proceso(cursor, proceso + ': ' + filename, pasos_proceso, paso, 'inserta en domicilios nuevos')
+        cf.logging_proceso(cursor, proceso + ': ' + filename, pasos_proceso, paso, 'inserta en domicilios nuevos')
        
         paso = 7
         staging_step_6 = "update Datos_generales.Domicilio a, Staging.tmp_solicitudes b set a.Calle = b.CALLE,"
@@ -241,7 +237,7 @@ if filename in files:
         staging_step_6 += " a.UltModificacion = b.FECHASOL"
         staging_step_6 += " where a.Num_cliente = b.numcte;"
         cursor.execute(staging_step_6)
-        logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso, paso,'Actualiza domicilios con existentes')
+        cf.logging_proceso(cursor,proceso + ': ' + filename,pasos_proceso, paso,'Actualiza domicilios con existentes')
 
 
         print('Proceso de carga terminado: ' + filename)
